requirements.txt
pyyaml
jsonschema
referencing